reporter: allow configuring the certificate and keyfile
Add a helper to generate a certificate while at it. The tool is intended
to create the dummy certificate. The ability to create a key for the
reporter service just exists for local development/testing purposes.

The mtime check should probably be replaced by fsnotify, but that would
add another dependency and is not too important at the moment.